requirements.txt
fastapi[all]>0.116,<0.117
alembic>1.16,<1.17
passlib[bcrypt]
pyjwt
flake8
black